chore: msw 폴더 위치 이전 & 테스트용 코드 제거
