+++ mocks/browser.ts
