+++ src/app/layout.tsx
@@ -1,42 +1,39 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
+import type {Metadata} from "next";
+import {Geist, Geist_Mono} from "next/font/google";
 import "./globals.css";
-import MSWProvider from "@/mocks/components/mswProvider";
+import MSWProvider from "./mswProvider";
 
-if (
-  process.env.NEXT_RUNTIME === "nodejs" &&
-  process.env.NODE_ENV === "development"
-) {
-  const { worker } = await import("@/mocks/server");
+if (process.env.NEXT_RUNTIME === "nodejs" && process.env.NODE_ENV === "development") {
+    const {worker} = await import("../../mocks/server");
 
-  worker.listen();
+    worker.listen();
 }
 
 const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
+    variable: "--font-geist-sans",
+    subsets: ["latin"]
 });
 
 const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
+    variable: "--font-geist-mono",
+    subsets: ["latin"]
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+    title: "Create Next App",
+    description: "Generated by create next app"
 };
 
 export default async function RootLayout({
-  children,
+    children
 }: Readonly<{
-  children: React.ReactNode;
+    children: React.ReactNode;
 }>) {
-  return (
-    <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>
-        <MSWProvider>{children}</MSWProvider>
-      </body>
-    </html>
-  );
+    return (
+        <html lang="en">
+            <body className={`${geistSans.variable} ${geistMono.variable}`}>
+                <MSWProvider>{children}</MSWProvider>
+            </body>
+        </html>
+    );
 }
