--- src/app/test.tsx
@@ -1,15 +1,0 @@
-"use client";
-
-const Test = () => {
-  const test = async () => {
-    const test = await fetch("https://api.example.com/example/user").then(
-      (res) => res.json()
-    );
-
-    console.log(test);
-  };
-
-  return <button onClick={test}>Test</button>;
-};
-
-export default Test;
